+++ src/maplibre/utils/index.js
@@ -1,2 +1,2 @@
-export { default as getSourceCoordinates } from './getSourceCoordinates';
 export { default as getMercatorResolution } from './getMercatorResolution';
+export { default as getSourceCoordinates } from './getSourceCoordinates';